Move image request handling out of main.go
